% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POUMM.R
\name{NewAbcPOUMMCppObject}
\alias{NewAbcPOUMMCppObject}
\title{Create an instance of the RCPP_PMM module for a given tree and trait data}
\usage{
NewAbcPOUMMCppObject(x, tree)
}
\arguments{
\item{x}{a numerical vector of size N, where N is the number of tips in tree}

\item{tree}{a phylo object}
}
\value{
an object to be passed as argument of the \link{POUMMLogLikCpp} function.
}
\description{
Create an instance of the RCPP_PMM module for a given tree and trait data
}
\seealso{
\code{\link{PMMLogLikCpp}}
}
