% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMM.R
\name{NewPMMCppObject}
\alias{NewPMMCppObject}
\title{Create an instance of the Rcpp module for a given tree and trait data}
\usage{
NewPMMCppObject(x, tree)
}
\arguments{
\item{x}{a numerical vector of size N, where N is the number of tips in tree}

\item{tree}{a phylo object}
}
\value{
an object to be passed as argument of the \link{POUMMLogLikCpp} function.
}
\description{
Create an instance of the Rcpp module for a given tree and trait data
}
\seealso{
\link{POUMMLogLikCpp}
}
