% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POUMM.R
\name{POUMMLogLik}
\alias{POUMMLogLik}
\title{Calculate the POUMM log-likelihood for a given tree, data and model parameters}
\usage{
POUMMLogLik(x, tree, x0, alpha, theta, sigma2, sigmae2,
  ord = reorder(tree, order = "postorder", index.only = TRUE))
}
\arguments{
\item{x}{a numerical vector of size N, where N is the number of tips in tree}

\item{tree}{a phylo object}

\item{x0, alpha, theta, sigma2, sigmae2}{parameters of the PMM:
\describe{
\item{x0}{value at the root of tree excluding whte noise;}
\item{alpha}{selection strength parameter of the OU process;}
\item{theta}{long-term optimum of the OU process;}
\item{sigma2}{unit-time variance increment of the heritable component (OU process);}
\item{sigmae2}{variance of the non-heritable component.}
}}

\item{ord}{integer vector: indices of the rows in tree$edge in pruning order. 

Default: \code{reorder(tree, order = "postorder", index.only = TRUE)}.}
}
\value{
the log-likelihood value.
}
\description{
Calculate the POUMM log-likelihood for a given tree, data and model parameters
}
