% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POUMM.R
\name{POUMMLogLikCpp}
\alias{POUMMLogLikCpp}
\title{Calculate the PMM log-likelihood for a given tree, data and model parameters using the RCPP_PMM module}
\usage{
POUMMLogLikCpp(x, tree, x0, alpha, theta, sigma2, sigmae2,
  cppObject = New3PointPOUMMCppObject(x, tree),
  mode = getOption("SPLITT.postorder.mode", 0))
}
\arguments{
\item{x}{a numerical vector of size N, where N is the number of tips in tree}

\item{tree}{a phylo object}

\item{x0}{parameters of the PMM:
\describe{
\item{x0}{value at the root of tree excluding whte noise;}
\item{alpha}{selection strength parameter of the OU process;}
\item{theta}{long-term optimum of the OU process;}
\item{sigma2}{unit-time variance increment of the heritable component (OU process);}
\item{sigmae2}{variance of the non-heritable component.}
}}

\item{alpha}{parameters of the PMM:
\describe{
\item{x0}{value at the root of tree excluding whte noise;}
\item{alpha}{selection strength parameter of the OU process;}
\item{theta}{long-term optimum of the OU process;}
\item{sigma2}{unit-time variance increment of the heritable component (OU process);}
\item{sigmae2}{variance of the non-heritable component.}
}}

\item{theta}{parameters of the PMM:
\describe{
\item{x0}{value at the root of tree excluding whte noise;}
\item{alpha}{selection strength parameter of the OU process;}
\item{theta}{long-term optimum of the OU process;}
\item{sigma2}{unit-time variance increment of the heritable component (OU process);}
\item{sigmae2}{variance of the non-heritable component.}
}}

\item{sigma2}{parameters of the PMM:
\describe{
\item{x0}{value at the root of tree excluding whte noise;}
\item{alpha}{selection strength parameter of the OU process;}
\item{theta}{long-term optimum of the OU process;}
\item{sigma2}{unit-time variance increment of the heritable component (OU process);}
\item{sigmae2}{variance of the non-heritable component.}
}}

\item{sigmae2}{parameters of the PMM:
\describe{
\item{x0}{value at the root of tree excluding whte noise;}
\item{alpha}{selection strength parameter of the OU process;}
\item{theta}{long-term optimum of the OU process;}
\item{sigma2}{unit-time variance increment of the heritable component (OU process);}
\item{sigmae2}{variance of the non-heritable component.}
}}

\item{cppObject}{a previously created object returned by either 
\code{\link{New3PointPOUMMCppObject}} or \code{\link{NewAbcPOUMMCppObject}} .}

\item{mode}{an integer denoting the mode for traversing the tree, i.e. serial vs parallel.}
}
\value{
the log-likelihood value.
}
\description{
Calculate the PMM log-likelihood for a given tree, data and model parameters using the RCPP_PMM module
}
