% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ThreePointUsingSPLITT__AbcPOUMM__AlgorithmType}
\alias{ThreePointUsingSPLITT__AbcPOUMM__AlgorithmType}
\alias{Rcpp_ThreePointUsingSPLITT__AbcPOUMM__AlgorithmType-class}
\title{\code{TraversalAlgorithm}-type used in \code{AbcPOUMM}}
\description{
\code{TraversalAlgorithm}-type used in \code{AbcPOUMM}
}
