% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ThreePointUsingSPLITT__AbcPOUMM__TraversalAlgorithm}
\alias{ThreePointUsingSPLITT__AbcPOUMM__TraversalAlgorithm}
\alias{Rcpp_ThreePointUsingSPLITT__AbcPOUMM__TraversalAlgorithm-class}
\title{Base class for \code{ThreePointUsingSPLITT::AbcPOUMM::AlgorithmType}}
\description{
Base class for \code{ThreePointUsingSPLITT::AbcPOUMM::AlgorithmType}
}
