% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ThreePointUsingSPLITT__ThreePointPMM__AlgorithmType}
\alias{ThreePointUsingSPLITT__ThreePointPMM__AlgorithmType}
\alias{Rcpp_ThreePointUsingSPLITT__ThreePointPMM__AlgorithmType-class}
\title{\code{TraversalAlgorithm}-type used in \code{ThreePointPMM}}
\description{
\code{TraversalAlgorithm}-type used in \code{ThreePointPMM}
}
