% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ThreePointUsingSPLITT__ThreePointPMM__TraversalAlgorithm}
\alias{ThreePointUsingSPLITT__ThreePointPMM__TraversalAlgorithm}
\alias{Rcpp_ThreePointUsingSPLITT__ThreePointPMM__TraversalAlgorithm-class}
\title{Base class for \code{ThreePointUsingSPLITT::ThreePointPMM::AlgorithmType}}
\description{
Base class for \code{ThreePointUsingSPLITT::ThreePointPMM::AlgorithmType}
}
