% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ThreePointUsingSPLITT__ThreePointPOUMM__AlgorithmType}
\alias{ThreePointUsingSPLITT__ThreePointPOUMM__AlgorithmType}
\alias{Rcpp_ThreePointUsingSPLITT__ThreePointPOUMM__AlgorithmType-class}
\title{\code{TraversalAlgorithm}-type used in \code{ThreePointPOUMM}}
\description{
\code{TraversalAlgorithm}-type used in \code{ThreePointPOUMM}
}
