% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ThreePointUsingSPLITT__ThreePointPOUMM__TraversalAlgorithm}
\alias{ThreePointUsingSPLITT__ThreePointPOUMM__TraversalAlgorithm}
\alias{Rcpp_ThreePointUsingSPLITT__ThreePointPOUMM__TraversalAlgorithm-class}
\title{Base class for \code{ThreePointUsingSPLITT::ThreePointPOUMM::AlgorithmType}}
\description{
Base class for \code{ThreePointUsingSPLITT::ThreePointPOUMM::AlgorithmType}
}
